+++ test/instrumentation/modules/ws.js
@@ -3,8 +3,14 @@
 var agent = require('../../..').start({
   serviceName: 'test',
   secretToken: 'test',
-  captureExceptions: false
+  captureExceptions: false,
+  metricsInterval: 0
 })
+
+var version = require('restify/package.json').version
+var semver = require('semver')
+
+if (semver.lt(process.version, '8.6.0') && semver.gte(version, '7.0.0')) process.exit()
 
 var test = require('tape')
 var WebSocket = require('ws')